% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributiondemo.R
\name{distributiondemo}
\alias{distributiondemo}
\title{Display Census data as a bar chart and table.}
\usage{
distributiondemo()
}
\description{
Simple shiny application to display distribution of count data.
}
\examples{
\dontrun{
}
}
\author{
Chris Hansen
}

