% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdemo.R
\name{mapdemo}
\alias{mapdemo}
\title{Display Census data on an interactive map.}
\usage{
mapdemo()
}
\description{
Simple shiny application to display Census data on a map.
}
\examples{
\dontrun{
}
}
\author{
Chris Hansen
}

