% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WARD2013}
\alias{WARD2013}
\title{Ward 2013}
\format{A \code{SpatialPolygonsDataFrame} containing columns:
\describe{
  \item{WARD2013}{Ward code.}
  \item{WARD013_name}{Ward label.}
}}
\source{
\url{http://www.stats.govt.nz/browse_for_stats/Maps_and_geography/Geographic-areas/digital-boundary-files.aspx}
}
\usage{
WARD2013
}
\description{
A \code{SpatialPolygonsDataFrame} containing the Ward 2013 pattern.  Features
have been reprojected using a web mercator projection (EPSG: 4326) for
convenient display in a web browser.
}
\details{
The ward features have been simplified using a 100m tolerance to try
and keep the overall size down.  Still, the feature class is still large
overall, so rendering inside a web browser might still be problematic.
}
\keyword{datasets}

